requirements.txt
Mako==0.7.3
MarkupSafe==0.15
bottle==0.11.4
py==1.4.12
pyparsing==1.5.6
pytest==2.3.4
python-dateutil==2.1
six==1.2.0

# wsgiref comes with python 3
#wsgiref==0.1.2

# SQLAlchemy 0.8.0 isn't in PyPI yet so get it from SourceForge
#SQLAlchemy==0.8.0
http://downloads.sourceforge.net/project/sqlalchemy/sqlalchemy/0.8.0b2/SQLAlchemy-0.8.0b2.tar.gz

# don't explicitly install distribute
#distribute==0.6.24

# Required for requests
chardet2
urllib3

# The requests 0.14.2 package is broken for Python 3 so use this temp package for now
requests-packaging-fixes
#requests==0.14.2